+++ R/estimateCIs.R
@@ -12,14 +12,14 @@
 #'
 #' @examples
 #' tbl <- data.frame(rho = c(0.5, 0.7, 0.3), se = c(0.1, 0.2, 0.05))
-#' calculate_CIs(tbl, rho_var = "rho", se_var = "se", method = "raykov")
+#' calculateCIs(tbl, rho_var = "rho", se_var = "se", method = "raykov")
 #'
 #' @export
-calculate_CIs <- function(tbl, rho_var, se_var, method = "raykov", conf_level = 0.95) {
+calculateCIs <- function(tbl, rho_var, se_var, method = "raykov", conf_level = 0.95) {
   # Load necessary packages
   require(rlang)
-  require(psych)
-  
+
+
   # Get the name of the rho column regardless of input type
  if(mode(rho_var) != "character") {
     stop("method must be a character string")
@@ -39,23 +39,23 @@ calculate_CIs <- function(tbl, rho_var, se_var, method = "raykov", conf_level = 
 
   # Create a copy of tbl to avoid modifying the original data frame
   tbl_out <- tbl
-  
+
   if (method == "raykov") {
     # Apply Fisher's r to z transform
-    z_vals <- fisherz(tbl_out[[rho_col_name]])
+    z_vals <- psych::fisherz(tbl_out[[rho_col_name]])
     se_vals <- tbl_out[[se_col_name]] / (1 - tbl_out[[rho_col_name]]^2)
-    
+
     tbl_out[[z_col_name]] <- z_vals
     tbl_out[[sez_col_name]] <- se_vals
-    tbl_out[[plusse_col_name]] <- fisherz2r(z_vals + qnorm((1 + conf_level) / 2) * se_vals)
-    tbl_out[[minusse_col_name]] <- fisherz2r(z_vals - qnorm((1 + conf_level) / 2) * se_vals)
-    
+    tbl_out[[plusse_col_name]] <- psych::fisherz2r(z_vals + qnorm((1 + conf_level) / 2) * se_vals)
+    tbl_out[[minusse_col_name]] <- psych::fisherz2r(z_vals - qnorm((1 + conf_level) / 2) * se_vals)
+
     # H0: r = 0. Compute z-test and p-values
     ztest_vals <- tbl_out[[rho_col_name]] / tbl_out[[se_col_name]]
     tbl_out[[ztest_col_name]] <- ztest_vals
     tbl_out[[p1tail_col_name]] <- pnorm(q = abs(ztest_vals), lower.tail = FALSE)
     tbl_out[[p2tail_col_name]] <- 2 * tbl_out[[p1tail_col_name]]
-    
+
   } else {
     # Set the multiplier for CI calculations based on the method
     ci_multi <- switch(method,
@@ -68,7 +68,7 @@ calculate_CIs <- function(tbl, rho_var, se_var, method = "raykov", conf_level = 
     tbl_out[[plusse_col_name]] <- tbl_out[[rho_col_name]] + ci_multi * tbl_out[[se_col_name]]
     tbl_out[[minusse_col_name]] <- tbl_out[[rho_col_name]] - ci_multi * tbl_out[[se_col_name]]
   }
-  
+
   return(tbl_out)
 }